Moved CLI to a subdirectory for future extensibility
